Replace common util with page objects
